new commit

+++ Mars Mips Simulator/Function.cs
@@ -173,7 +173,7 @@ namespace Mars_Mips_Simulator
             int number;
             if (int.TryParse(num2, System.Globalization.NumberStyles.HexNumber, null, out number))
             {
-                return (Convert.ToInt32(num1, 16) * (int.Parse(num2))).ToString();
+                return (Convert.ToInt64(num1, 16) * (int.Parse(num2))).ToString();
 
             }
             else if (int.TryParse(num2, System.Globalization.NumberStyles.HexNumber, null, out number) && int.TryParse(num1, System.Globalization.NumberStyles.HexNumber, null, out number))
@@ -183,7 +183,7 @@ namespace Mars_Mips_Simulator
             else
             {
 
-                return (Convert.ToInt32(num2, 16) * Convert.ToInt32(num1, 16)).ToString();
+                return (Convert.ToInt64(num2, 16) * Convert.ToInt32(num1, 16)).ToString();
             }
